git-svn-id: https://svn.neo4j.org/components/neo/trunk@2846 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ enterprise/src/java/org/neo4j/impl/core/NodeManager.java
@@ -713,12 +713,12 @@ public class NodeManager
         }
     }
 
-    public int getHighestPossibleIdInUse( Class<?> clazz )
+    public long getHighestPossibleIdInUse( Class<?> clazz )
     {
         return idGenerator.getHighestPossibleIdInUse( clazz );
     }
 
-    public int getNumberOfIdsInUse( Class<?> clazz )
+    public long getNumberOfIdsInUse( Class<?> clazz )
     {
         return idGenerator.getNumberOfIdsInUse( clazz );
     }
